+++ R/select-helpers.R
@@ -8,6 +8,7 @@
 #' * `num_range()`: a numerical range like x01, x02, x03.
 #' * `one_of()`: variables in character vector.
 #' * `everything()`: all variables.
+#' * `last_var()`: last variable, possibly with an offset.
 #'
 #' @param match A string.
 #' @param ignore.case If `TRUE`, the default, ignores case when matching
@@ -25,6 +26,8 @@
 #' vars_select(nms, matches(".t."))
 #' vars_select(nms, Petal.Length, Petal.Width)
 #' vars_select(nms, everything())
+#' vars_select(nms, last_var())
+#' vars_select(nms, last_var(offset = 2))
 #'
 #' vars <- c("Petal.Length", "Petal.Width")
 #' vars_select(nms, one_of(vars))
@@ -111,6 +114,23 @@ one_of <- function(..., .vars = peek_vars()) {
 #' @rdname select_helpers
 everything <- function(vars = peek_vars()) {
   seq_along(vars)
+}
+
+
+#' @export
+#' @param offset Set it to `n` to select the nth var from the end.
+#' @rdname select_helpers
+last_var <- function(offset = 0, vars = peek_vars()) {
+  stopifnot(is_integerish(offset))
+  n <- length(vars)
+
+  if (offset && n <= offset) {
+    abort("`offset` must be smaller than the number of variables")
+  } else if (n == 0) {
+    vars
+  } else {
+    vars[[n - offset]]
+  }
 }
 
 match_vars <- function(needle, haystack) {