+++ tests/testthat/test-select-helpers.R
@@ -220,3 +220,12 @@ test_that("can select with .data pronoun (#2715)", {
   expect_identical(vars_select(c("a", "b", "c"), .data$a : .data$b), c(a = "a", b = "b"))
   expect_identical(vars_select(c("a", "b", "c"), .data[["a"]] : .data[["b"]]), c(a = "a", b = "b"))
 })
+
+test_that("last_var() selects last argument with offset", {
+  vars <- letters[1:3]
+  expect_identical(last_var(0, vars), "c")
+  expect_identical(last_var(2, vars), "a")
+
+  expect_error(last_var(3, vars), "`offset` must be smaller")
+  expect_identical(last_var(vars = chr()), chr())
+})